ENH: Generate pure plug mask program file
An ITK filter is added to evaluate a sample set to check whether they
belong to a pure plug or not within the defined threshold.
A Mahalanobis-weighted Euclidean distance is used as the metric.
This membership filter is called in a program that collects samples as
normalized intensity values (between 0 and 1) of the input multimodal images.
Note that it is assumed that input multimodal scans are aligned in
physical space.

Usage:
$ bin/GeneratePurePlugMask \
  --inputImageModalities t1,t2,IDWI \
  --threshold 0.2 \
  --numberOfSubSamples 2,2,2 \
  --outputMaskFile binary_pure_mask.nrrd